Вроде починено редактирование лаб (создание будет тестироваться ещё). Исправлено представление категорий (не ругается на пустой список их). Подправлена отправка действий студента (при завершении лабы).
git-svn-id: https://svtz.ru/svn/graphlabs.site/trunk@341 e8704808-1a2a-453a-8987-66899a109cbe

+++ GraphLabs.Site/Controllers/LabsController.cs
@@ -5,13 +5,16 @@ using Newtonsoft.Json;
 using System;
 using System.Web.Mvc;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using GraphLabs.DomainModel;
 using GraphLabs.DomainModel.Contexts;
 using GraphLabs.DomainModel.Repositories;
 using GraphLabs.Site.Models.CreateLab;
 using GraphLabs.Site.Models.Infrastructure;
 using GraphLabs.Site.Models.Lab;
+using GraphLabs.Site.Utils;
 
 namespace GraphLabs.Site.Controllers
 {
@@ -70,10 +73,12 @@ namespace GraphLabs.Site.Controllers
                 return Json(new JSONResultCreateLab( ResponseConstants.LabWorkExistErrorSystemName, Name ));
 			};
 
-			LabWork lab = _labWorksContext.LabWorks.CreateNew();
-			lab.Name = Name;
-            lab.AcquaintanceFrom = DateTime.Parse(DateFrom); // ParseDate.Parse(DateFrom);
-			lab.AcquaintanceTill = DateTime.Parse(DateTill);
+            if (!ParseDate.TryParse(DateFrom) || !ParseDate.TryParse(DateTill)) return Json(new JSONResultCreateLab( ResponseConstants.LabWorkDateCouldNotBeParsed, Name));
+            LabWork lab = _labWorksContext.LabWorks.CreateNew();
+            var message = "";
+            lab.Name = Name;
+            lab.AcquaintanceFrom = ParseDate.Parse(DateFrom, out message); 
+			lab.AcquaintanceTill = ParseDate.Parse(DateTill, out message);
             //_changesTracker.SaveChanges();
 			_labRepository.SaveLabEntries(lab.Id, JsonConvert.DeserializeObject<long[]>(JsonArr));
 			_labRepository.DeleteExcessTaskVariantsFromLabVariants(lab.Id);
@@ -88,11 +93,12 @@ namespace GraphLabs.Site.Controllers
 			{
 				return Json(new JSONResultCreateLab(ResponseConstants.LabWorkExistErrorSystemName, Name));
 			};
-
-			LabWork lab = _labRepository.GetLabWorkById(id);
+            if (!ParseDate.TryParse(DateFrom) || !ParseDate.TryParse(DateTill)) return Json(new JSONResultCreateLab(ResponseConstants.LabWorkDateCouldNotBeParsed, Name));
+            LabWork lab = _labRepository.GetLabWorkById(id);
+            var message = "";
 			lab.Name = Name;
-			lab.AcquaintanceFrom = DateTime.Parse(DateFrom);
-			lab.AcquaintanceTill = DateTime.Parse(DateTill);
+			lab.AcquaintanceFrom = ParseDate.Parse(DateFrom, out message);
+			lab.AcquaintanceTill = ParseDate.Parse(DateTill, out message);
 
 			_labRepository.DeleteEntries(id);
 			lab.LabEntries.Clear();
